refactor: apply changes to use esm module
